table functions moved to separate file
