+++ Core/wfmFunctions.py
@@ -14,19 +14,19 @@ from Util import dict_map
 from math import *
 from LogConfig import *
 
-def store_wf(event, table, WF):
-    """
-    Store a wavform in a table
-    """
-    row = table.row
-    for isens,wf in WF.items():
-        for t,e in zip(wf.time_mus, wf.ene_pes):
-            row['event'] = event
-            row['ID'] = isens
-            row['time_mus'] = t
-            row['ene_pes'] = e
-            row.append()
-    table.flush()
+# def store_wf(event, table, WF):
+#     """
+#     Store a wavform in a table
+#     """
+#     row = table.row
+#     for isens,wf in WF.items():
+#         for t,e in zip(wf.time_mus, wf.ene_pes):
+#             row['event'] = event
+#             row['ID'] = isens
+#             row['time_mus'] = t
+#             row['ene_pes'] = e
+#             row.append()
+#     table.flush()
 
 # def read_twf(pmttwf, event_number):
 #     """
@@ -76,26 +76,26 @@ def store_wf(event, table, WF):
 #
 #     return pd.Panel(sensors)
 
-def read_wf(table,event_number,isens):
-    '''
-        Reads table and returns the waveform (time_mus and ene_pes) corresponding
-        to sensor isens of event event_number.
-    '''
-    try:
-        return zip(*[ (row['time_mus'],row['ene_pes']) for row in table.iterrows() if row['event']== event_number and row['ID']== isens])
-    except:
-        logger.error('[read_wf]: empty sensor found: {}'.format(isens))
-
-
-def read_wf_table( table, event_number ):
-    """
-    Reads back the TWF of the PMTs/SiPMs for event number:
-    input: the twf table of the PMTs,(SiPMs) a list with the PMT (SiPMs) indexes and the event number
-    outputs: a PMT/SiPM panel
-
-    """
-    sensor_list = set(table.read_where('event == {}'.format(event_number),field='ID'))
-    return pd.Panel({ isens : wf2df(*read_wf(table,event_number,isens)) for isens in sensor_list})
+# def read_wf(table,event_number,isens):
+#     '''
+#         Reads table and returns the waveform (time_mus and ene_pes) corresponding
+#         to sensor isens of event event_number.
+#     '''
+#     try:
+#         return zip(*[ (row['time_mus'],row['ene_pes']) for row in table.iterrows() if row['event']== event_number and row['ID']== isens])
+#     except:
+#         logger.error('[read_wf]: empty sensor found: {}'.format(isens))
+#
+#
+# def read_wf_table( table, event_number ):
+#     """
+#     Reads back the TWF of the PMTs/SiPMs for event number:
+#     input: the twf table of the PMTs,(SiPMs) a list with the PMT (SiPMs) indexes and the event number
+#     outputs: a PMT/SiPM panel
+#
+#     """
+#     sensor_list = set(table.read_where('event == {}'.format(event_number),field='ID'))
+#     return pd.Panel({ isens : wf2df(*read_wf(table,event_number,isens)) for isens in sensor_list})
 
 
 def rebin_twf(t, e, stride = 40):
